fix: key prop in table
