+++ dashboard/app/components/stats/types.ts
@@ -1,0 +1,37 @@
+interface DataRow {
+	// Common
+	visitors?: number;
+	visitors_percentage?: number;
+	// Mixed
+	path?: string;
+	bounces?: number;
+	bounce_rate?: number;
+	duration?: number;
+	// Pages
+	pageviews?: number;
+	pageviews_percentage?: number;
+	// Duration
+	duration_upper_quartile?: number;
+	duration_lower_quartile?: number;
+	duration_percentage?: number;
+	// Referrers
+	referrer?: string;
+	// Sources
+	source?: string;
+	// Mediums
+	medium?: string;
+	// Campaigns
+	campaign?: string;
+	// Browsers
+	browser?: string;
+	// Operating Systems
+	os?: string;
+	// Devices
+	device?: string;
+	// Countries
+	country?: string;
+	// Languages
+	language?: string;
+}
+
+export type { DataRow };
