+++ dashboard/app/routes/$hostname.$query.tsx
@@ -1,12 +1,14 @@
 import {
 	json,
-	type MetaFunction,
-	type ClientLoaderFunctionArgs,
 	useLoaderData,
 	useParams,
+	type ClientLoaderFunctionArgs,
+	type MetaFunction,
 } from '@remix-run/react';
+
 import { StatsTable } from '@/components/stats/StatsTable';
-import { type DatasetItem, fetchStats, isDatasetItem } from '@/utils/stats';
+import type { DataRow } from '@/components/stats/types';
+import { fetchStats, isDatasetItem, type DatasetItem } from '@/utils/stats';
 
 export const meta: MetaFunction = () => {
 	return [{ title: 'Dashboard | Medama' }];
@@ -33,11 +35,16 @@ export default function Index() {
 	// We can safely assume that the dataset items are present as the loader function
 	// has already validated the query parameter
 	const query = params.query as keyof typeof data;
-	const stats = data[query];
+	let stats = data[query] as DataRow[];
 
 	if (!stats || typeof stats === 'number') {
 		return <div>No data found.</div>;
 	}
+
+	// Add id to each stat item which is the index of the item in the array
+	stats = Array.isArray(stats)
+		? stats.map((stat, index) => ({ ...stat, id: index }))
+		: [stats];
 
 	return <StatsTable query={query} data={stats} />;
 }
